Refactor syncrhonize method
Take return out of transaction block. Reduce amount of processing performed in transaction. Not necessary to ensure read is done in the transaction because there is an inherent delay between the user reading the confirmation checklist, and pressing the okay button. Any changes to the database during that time will be synced too, so don't worry about the small window between reading the entries and syncing them.