Added sync_now method.
This method can be used to ensure automated tasks that do not need user confirmation are synchronized to the production server immediately, if possible. If not, the changes are held in a commit so they are all synced at the same time whenever the record that prevented immediate sync is itself synced.