Improve multithreaded sync tests
Tests now take into account that possibility of a thread sleeping because it is pre-empted by another thread, and not just sleeping at the explicit `sleep` instructions in the test. A possible false positive exists on line 260 because there is no way to know if Thread 2 is sleeping because it was pre-empted, or sleeping because it's waiting for the lock on `source_record`, so we assume that it is the latter and allow thread 1 to wake. Technically we do not know if thread 2 has attempted to read `source_record` yet.