Fix rake not creating triggers
If the test environment used only a single connection, and the rake task was run in that environment, the extended version of create_table would not create triggers because it found that it was migrating the production database (which was also the staging database). Now if there is a single connection, we ignore the fact that we're in the production database and still create triggers.