Simplify synced? method
No need to load all the records via a has_many. The existence of one CommitEntry is enough to indicate the record has unsynced changes so a has_one will suffice. Removed the option to ask for only contained entries when checking sync status. The initial idea was that the UI could hide changes that were going to be autosynced, but this is probably not worth creating a second "contained_stagehand_sync_indicator" association for.