Fix bug where subject was being preconfirmed in sync_now
It must not preconfirm changes to the subject record because sync_now is for use in automated tasks that should be immediately synced unless they conflict with existing changes.

This bug was caused by the change to checklist where the subject record did not need to be preconfirmed since it was implied by the fact that we're seeing what else needs to be confirmed in order to sync the subject record.