Improve syncing performance of millions of records
An index on operation and commit_id was added to make not_in_progress scope faster when millions of commit entries are present.

Also simplify the autosyncable_entries method by not requring it to return only the latest entry for any given key. This avoids the need to group entries, which was an expensive operation when the commit entries table was large. Instead, uniqueness is handled in ruby, and only when actually required.