Make it possible to use connected_to_x? in a migration
Because not all migrations should execute on both staging and production, there needs to be a way to skip execution if the migration is running on a particular database. This modification to the migrate rake task uses Database.with_connection instead of ActiveRecord::Base.establish_connection so the current connection is monitored and can be read by Database.connected_to_staging? and Database.connected_to_production?

This change also ensures that the database connection is restored when finished. Without this, there was the potential to affect subsequent operations should the migrate task be run from within a ruby script instead of as a one-off task from the command line.