Cleaning up the code
Since the Migrate/Rollback use the same code to run on both databases, I pulled out the dup code and just pass in the unique part as a block.

Also found a bug in my code, where we are calling rollback twice. This is because enhance runs both methods, so I was calling rollback twice. Found that you can overwrite a task with .clear so that it clears the existing prerequisites and actions of a rake task, then call enhance to run the task we really want.