Creating Tests Migrations
Tests that on migrate/rollback that the command is in fact run on both databases.

Created fake dummy migration, this means so needed to add Rails.root to tasks so it would get the dummy/db/migrate when testing.

Added version:0 to the test schema.rb

Also since I am comparing the database versions, I created methods in database.rb to check what versions the production/staging databases we are on.  Will update synchronizer in next commit.