Fix bug where commits could appear to be empty
During intialization, if the database returned a commit's control entries in end, start instead of start, end, the id bounds for contained entries would be reversed. This led to a query that excluded all entries. Fixed by ensuring an order clause is included. The fix actually uses a reorder because the test applies an order scope to simulate the database returning records in opposite order.

Closes #6.