Improve sync performance
Instead of using the `session` column to shard commit entries for processing, just use the id. This grants more control over the theoretical maximum number of query results because it avoids the issue of a single session possibly containing thousands of entries. Performance gains on older versions of MySQL (5.1) are dramatic.