Optimize Auto Syncing
This commit changes the auto sync algorithm to avoid iterating through empty id ranges as we work our way from the max id to the min id. Now we return blocks of 1000 entries until we run out of entries. This ends up being approximately 3x faster when run on a real set of ~650,000 entries.