Fix mismatched_records returning identical records.
We were using LIMIT and OFFSET to grab batches of a fixed size and compare them. The issue is that if there is a difference in the number of records in each database, identical records in the two databases would appear in different batches.

We now grab records based on id range to ensure that identical records are always in the same batch.