Fixes a bug that could result in destroyed commit logs
We need to be certain we’re not deleting matching entries that belong to active commits. We can achieve this if we make a list of the entries in active commits before we start syncing, and only delete matching records that are not in that list, and that did not get created after the current entry being synced. Although simpler, for performance reasons we do not use the `not_in_progress` scope to limit the matching records as it makes the process more than 12x slower.