Hardcode staging connection name
Since Rails always connects to the database connection that matches the current environment, there’s no way to remove that database.yml entry even if the `staging_connection_name` is customized to something else.

This also has the potential to cause unexpected behaviour elsewhere in the middleware stack in situations where the `staging_connection_name` is different from the `Rails.env` value because Stagehand will swap database connections for all requests.

This specific situation caused a problem when Stagehand was used in conjunction with the Rehearsal gem. Rehearsal starts a transaction before the swap occurs, using the default Rails connection for the current environment, in this case, `production`. Rails then swaps the database to the staging connection, writes to the database. These writes are never rolled back by the Rehearsal gem because the transaction was started on a different connection than the one that was used to make changes.

The problem is exacerbated if the production connection name is the `Rails.env` value, as all edits not performed explicitly in the staging database will attempt to run on the production database (though they will raise `UnsyncedProductionWrite` exceptions).