Subject commit entries are no longer filtered
Fixed a bug where subject entries on a `Checklist` would be filtered by the relation filter if they matched. This had the effect of rejecting entries even though they were not being used to find related commits, but were actually the subject entries used to generate the `Checklist` in the first place.

To solve this problem we accumulate the subject entries and the commits those entries are a part of, and find the entries related to those. Previously we had used the `Checklist::related_entries` method to accumulate those subject and subject commit entries, which has the effect of rejecting any that do not pass the relation filter.