improve handling of records with serialized attributes
`CommitEntry#record` did not work for delete operations of  records with serialized attributes because the production record loads raw attribute values, not type-cast ones. To avoid this issue, we directly access the `AttributeSet` of the record we’re attempting to rebuild, and use the `write_from_database` method to write raw values.

Also fixed is a bug where we could not infer the STI subclass of a delete operation record since we were only looking in the staging database for the record’s inheritance column information. We now look at the production record for that information in the case of a delete operation.