Threadsafe update
This is a bit of a WIP but it allows stagehand to operate on Webbrick without `Rack::Lock`. The changes are geared towards ensuring that multi-threaded serving of requests does not cause the connection to be lost while connection swapping occurs in separate threads. The solution appears to be not to call `establish_connection` except during initialization. `establish_connection` removes the old connection pool, which causes a simultaneously executing thread to choke as it looks for that connection pool. Instead, we change the variable that contains the name of the connection pool we’re looking up. This requries some gymnastics, because while Rails 5 allows the name to change dynamically, it isn’t expecting separate threads to want the same `ActiveRecord` class to use separate connections.