Share a connection pool for a class across threads
While we want to be able to override a class’s connection on a per thread basis, sometimes we want to just set the connection once and have it apply to all threads. For instance, `StagingProbe` and `ProductionProbe` will never swap connections, and we don’t want to have to specify the connection in every thread. To avoid having to do this we make use of the original `@connection_specification_name` variable that we wrapped in a per-thread implementation, and use it as a fallback for the class in any thread that hasn’t had a connection specified. This allows us to use `establish_connection` and `remove_connection` as we would normally do when we don’t want a class to have per-thread connection spec names.