Make it easier to fake multi-database mode in tests
In some host app tests we stub which database we’re connected to in an attempt to simulate multi-database mode, while only using a single database. Since we’re using the same db name for both databases, we need to check to see if we’re connected to the staging database in order to lift the readonly flag.