Don’t establish or remove connection anymore
In order to be thread-safe, we no longer use establish and remove connection. This commit updates the test suite to use `connection_specification_name` swapping instead of direct calls to `establish_connection` and `remove_connection`.

We now use the ’primary’ connection pool when connected to staging. This avoids using a separate connection than the default connection when connected to the default database (staging).