Fix syncing records that violate column constraint
If two records would violate a column constraint if they exist simultanesouly are synced after one is deleted and the other subsequently created (or updated to violate the constraint), the sync would fail because the deleted record was not always synced first.

To avoid this issue, we ensure that records are synced deletes, then updates, then inserts, ensuring the record with the constrained value is removed before the new record with the same constrained value is updated or inserted.

Closes https://github.com/culturecode/stagehand/issues/48.