Don’t marke deleted production records as destroyed
If they are marked as destroyed, it causes url generation to treat them as unsaved records, making it impossible to generate a url to the live record. Instead, we can “tag” them with a dummy module `Stagehand::DeletedRecord`, allowing introspection like `record.is_a?(Stagehand::DeletedRecord)` in place of `record.destroyed?`.