Fix `record_class` infering for deleted records
If we ask a `CommitEntry` what its `record_class` is, it looks that information up by finding the record in the database so it can determine if it is a specific STI subclass. This works for all `CommitEntry` operations except `delete`. In that case we do not have a record in the database to look up. We would then look up the record in the production database where we expect it to reside since it hasn’t been deleted there yet. However, if the record has never been synced the lookup would return `nil` and we would explode.

Instead, we now return the base class if we cannot determine the STI subclass from the production database.