Fix bug where the wrong connection would be used
If haven’t set the connection in a thread, and then make a new thread where we set the connection to production, the original thread classes will begin using the production connection. This is an unintended side-effect of the attempt to not have to set `ProductionProbe` and `StagingProbe` connections in every thread, see bcaa412e4ecfdccaa0eef1b9d2bbc82c1e35be24.

Instead, we do not share connections under default circumstances and classes can now opt-into sharing (essentially opting out of the per-thread connection tracking behaviour that Stagehand adds)