Fix bug where transaction rollback broke commit
If a transaction surrounding a commit rolled back, there was potential for the commit’s start operation to be rolled back but an end operation still be created.

Instead, we check during the `end_commit` method if the commit was rolled back, in which case we log a warning and do not create an end operation.