Fix potential deadlock
The deadlock occurred in production when a user was syncing a checklist while another process was ending a commit. Process 1 started a commit. Process 2 created commit entries while in a transaction thereby acquiring a lock on those rows. Process 1 then ended the commit but was forced to wait for a write lock on the rows between its start and end entries in order to set the `commit_id` because Process 2 had a lock on some of those rows. Process 2 then attempted to delete its entries, but was forced to wait for Process 1 which had a range lock that included the rows. Hence the deadlock.

Instead of performing an update with `BETWEEN` to set the `commit_id` on entries in our commit, we perform a read to get those ids. A read lock has a lower priority than a write lock, so it will not block another process from deleting rows that are covered by that `BETWEEN`. Once we have the ids, we perform a write to update them, and since we only have the records that are part of our commit, we don’t run the risk of a deadlock if another process manipulates other entries between our start and end entries as we would have if updating with a range lock.