Allow cross database joins for staging models.
We now keep track of all staging model table names and append the DB name when referencing them on the live side. This allows us to do cross-database joins in MySQL. This is useful because it allows us to stop copying tables just to be able to perform joins on the live site.