Disable table name prefixing when production writes are enabled
When `Connection.allow_unsynced_production_writes` is `true`, we don’t want to force a staging model to use the staging database because we are explicitly trying to do production writes. The most obvious place that this applies is during database migrations. Without this fix, migrations to staging tables will fail since we end up trying to add the same column to the staging database twice.