Fix bug that incorrectly copied JSON columns
Using `connection.select_one` does not properly typecast JSON columns to Ruby hashes when read, leaving them as a JSON string. This meant that when we wrote the value to the Production database, the JSON object became a string of JSON stored as JSON.

Instead of the raw connection we can use an ActiveRecord object, specifically our `StagingProbe`, to read the data from the Staging database and automatically apply the correct typecasting. This ensures the JSON object is converted to a hash when we read it, and then back to a JSON object (instead of a string) when it is written to the Production database.

A slight complication arose when I subclassed `StagingProbe` to read from the staging database. Firstly, in sync operations where we’re in a transaction, changes made within that transaction were not copied to the Production database because the `StagingProbe` held its own connection separate from the one the transaction was occurring on. To avoid this, we ensure that not only `Staging::Model`, but the `StagingProbe` also uses the standard `ActiveRecord::Base` connection when already connected to the Staging database, ensuring it’s inside the sync transaction and can read changes made there.

Another issue arose where callbacks were not being run on sync. By assigning the `StagingRecordReader` table name to the table that we wanted to read from, that class was being detected as the record class for the `CommitEntry` we were syncing. This meant that instead of detecting the correct model based by matching the `table_name` column in the `CommitEntry`, we were detecting our probe that had been pointed at that table. This was solved by ignoring all subclasses of `Stagehand::Database::Probe` when detecting a `CommitEntry`’s record class.