Remove dependence on session column
By setting a session variable before we start recording entries for a commit we can ensure that entries are added with the correct `commit_id` instead of having to set the `commit_id` at the end of the commit. This in turn avoids issues where unexpected program termination could leave the commit entries in a state where the commit was not completed and its commits did not have a `commit_id`, making it impossible to know whether they are part of a commit or not.