Fix bug that allowed checklist to be incorrect
The checklist expands its list of affected entries to include all uncontained entries that match the spidered entries as a final step. During this expansion, we only apply the `uncontained` scope. However, this expansion is liable to find entries that match and are uncontained because they are part of a commit that is in progress.

Instead, we scope by `not_in_progress` to make sure sure these commit entries don't get included.