Use thread_variable_get/set
Instead of using a long-lived object to store a map of stacks per thread. Store the map on the thread so it dies with the thread. Using `Thread.current.thread_variable_set/get ` ensures that even in a Fiber, we use the same stack as the thread is using. See https://github.com/rails/rails/issues/7746.