Fix Model.quoted_table_name not being prefixed
`ActiveRecord::Base::quoted_table_name` internally caches its result. If we call this method we could get an unexpected result since we expect the value to change under different circumstances.

Instead, we override the method to check if the quoted table name needs a database prefix, and cache that result separately from the non-prefixed version.