Don't add stagehand to production
This method can be called in a migration to add stagehand to a table that didn't have it before, but since we run the migration on staging and production databases, it adds it to both unless the user is careful to only run it on staging.

Since there's no reason to add it to the production database, we no-op that case.