Expose method for detecting if table has triggers
Currently there's no support for table names that are longer than 40 characters because the trigger prefix + table name would be over the 64 character limit for identifiers. Because of this, removing Stagehand or even asking if a table has Stagehand will cause MySQL to throw an exception.

Instead, we make a method that can tell us if a table has Stagehand, allowing for use of the per table version of `remove_stagehand!` on only tables that are confirmed to have Stagehand. In the future we may want to support long table names by creating unique, fixed-length trigger names, but for now this is the workaround.