Check if table has triggers before removal
This makes it easier for developers to use the `remove_stagehand!` method when tables with long names exist. Those names that would be too long to make into a trigger name currently explode Stagehand when we try to remove or create a trigger. Since these tables will not have Stagehand triggers, there's no need to process them.

In the future we may want to add support for triggers on tables with long table names, but for now we add this quality of life improvement.