Expunge Commit Entries on remove_stagehand!
When removing stagehand from a table, we generally want to remove all associated commit entries. We do already enhance the `drop_table` method with this behaviour, however, if `remove_stagehand!` is called manually before dropping the table, we would not end up deleting the commit entries for that table.

Instead, we make the cleanup process part of `remove_stagehand!`.