Don’t allow early return from commit
If a commit block contained a `return`, it would break out of the `capture` method before the commit ended. This could leave commits without `end` operation entries.

Instead, we use `ensure` to always create an end entry unless there was an error in stagehand, or if the commit was rolled back.