Discard empty commit captures
When capturing a commit, we typically don’t want to keep it around if it contains no entries. Doing so can lead to commits that will never be synced if they are empty and have no subject.

Instead, we discard empty commits by default. An option can be passed to allow creation of an empty capture.