Fix bug where Exception exceptions were swallowed
`rescue` handles `StandardError` and its subclasses by default. This means if an `Exception` that does not descend from `StandardError` is raised in the capture block, we’ll swallow it by returning a value from the `ensure`.

While it’s generally discouraged to rescue `Exception`, in this case we immediately re-raise the exception. All we want is to record the existence of the exception so we don’t return a value from the `ensure` block.