Add sync_now! method
It is becoming a common case to want to immediately sync changes that take place in an automated process regardless of whether it will cause changes from another commit to also be synced. To simplify this procedure, a new `sync_now!` method has been added which will behave the same as a `sync_now` block except that it will not check to see if the `Checklist` requires confirmation before executing the sync.