Fix query generation for associations
Staging Models prefix their table name with the staging database when we're connected to production, but this was not always working when loading an association that uses a Staging Model. If we load that association while connected to staging, then connect to production and load that association again, the query will not prefix the Staging Model's table name as expected.

It turns out there is a cache on each model that stores the association's query AST, including the table name. This means that it does not recalculate the table name when swapping connections.

In order to resolve this we can maintain two separate caches, one for staging and one for production. I opted to go this route instead of the clearing the cache whenever we swap connections so we can still benefit from the caching. I also opted not to just always include the database name prefix on queries because the presence of the prefix is also controlled by whether we're allowing production writes. It seemed safer to solve the issue instead of remembering that we can't do any lookups of Staging Models via associations when in a production write block. And lastly I opted not to solve the issue by clearing the cache when in a production write block because the auto sync task is constantly writing to the production database using that block and we don't want ot continually be clearing the cache.