Fix auto syncing record with foreign key constraint
We were exploding when auto syncing a record with a foreign key constraint because we process commit entries from latest to earliest in order to avoid performing unnecessary syncs of records that have been destroyed. However, by syncing in reverse order we end up attempting to delete a record with a foreign key constraint before the record it depends on.

Instead, we sync in ascending order so the records are deleted from the production database in the same order as they were in the staging database. This means we must look forward instead of backwards when deleting matching commit entries which requires that we avoid deleting a delete entry that was added after the list of autosyncable entries was determined, but before we look for entries matching the one we just synced in order to remove those.