Don't sync in-progress control entries
When we generate a checklist we need to avoid including the start operation commit entry if that commit is in-progress, even if the subject matches the entries we're adding to the checklist. Failing to do so would allow a sync that happens during another commit to destroy that commit's start operation if its subject matches a record we're syncing.