perf: Optimize autosync list generation
We are looking for commit entries that are not part of a commit by performing an anti-join. As part of the left outer join the results are distincted, however it should be unnecessary to do this since we're only going to be considering rows where the `record_id` is null. If a row would have been removed by distincting then it would mean it would have had to have a `record_id` and would have been discarded anyway.

By removing the distinct we see a 10x speedup in query speed to get the list of autosyncable records.