fix: Create specs to error when trying to sync
We remove the IndexPage::Documents subtype from Pages. But when we deleted the corresponding records, it exploding when trying to sync. This is because the subtype no longer existed, so when we tried to determine the record_class name, it exploded with a name error.

Instead of exploding, allow invalid STI subtypes to be deleted from the production database. In a similar way we are rescuing when we cannot infer the base_class name, we raise an IndeterminateRecordClass which will in turn, create a dummy class using the table_name. This way we can ensure that the production records are destroyed even if their subtype no longer exists.