Refactor captured flag to committed flag
Instead of flagging a commit entry as being captured, we flag the commit entry as having been committed. This simplifies the logic because we only have to set the flag at the end of the commit. Up to that point we can treat any commit entry with a commit_id as being in-progress. This eliminates the need to look at the whether we're capturing in SQL in order to set the flag.