Be defensive regarding in-progress entries
It is possible that after determining which entries are autosyncable, but before processing an entry, a matching entry is committed on a separate connection that is a match.

To avoid having to think about the possible race conditions, we apply the `uncontained` scope to the deletion scope.

Similarly we also need to add an uncontained scope when we're running in ghost mode to ensure the functionality is not changed from before, so we ensure the `not_in_progress` scope is still applied. We could have removed the scope and the tests would pass but we don't want to allow entries to be synced while the commit is in progress lest we have a handle in memory to something that was synced out from under us. We'll just autosync them on the next go round.